Organise into packages, command, task, exception Include private modifiers for attributes
