--- src/main/java/Main.java
@@ -1,25 +1,0 @@
-import java.util.ArrayList;
-import java.util.Scanner;
-
-public class Main {
-    public static void main(String[] args) throws IndexOutOfBoundsException {
-//        Scanner sc = new Scanner(System.in);
-        Lulu lulu = new Lulu("./data/lulu.txt");
-        lulu.run();
-        //lulu.greet();
-//        if (Storage.isSavePresent()) {
-//            Storage.readSave(lulu);
-//        }
-
-//        while (sc.hasNext()) {
-//            String s = sc.nextLine();
-//            String[] command = s.split(" ");
-//            if (command[0].compareTo("bye") == 0) {
-//                lulu.exit();
-//                break;
-//            }
-//            lulu.run(s);
-        }
-    }
-
-
