+++ src/main/java/lulu/Main.java
@@ -1,0 +1,9 @@
+package lulu;
+public class Main {
+    public static void main(String[] args) throws IndexOutOfBoundsException {
+        Lulu lulu = new Lulu("./data/lulu.txt");
+        lulu.run();
+    }
+}
+
+
