+++ src/main/java/lulu/command/Command.java
@@ -1,3 +1,8 @@
+package lulu.command;
+
+import lulu.TaskList;
+import lulu.UI;
+import lulu.Storage;
 public abstract class Command {
     public abstract void execute(TaskList tasks, UI ui, Storage storage);
