+++ src/main/java/lulu/command/TodoCommand.java
@@ -1,5 +1,15 @@
+package lulu.command;
+
+import lulu.TaskList;
+import lulu.UI;
+import lulu.Storage;
+
+import lulu.exception.LuluException;
+
+import lulu.task.Task;
+import lulu.task.Todo;
 public class TodoCommand extends Command {
-    String description;
+    private String description;
     public TodoCommand(String rest) throws LuluException {
         if (rest.isEmpty()) {
             throw new LuluException("(=✖ ᆺ ✖=) The description of a todo cannot be empty.");
