+++ src/main/java/lulu/exception/InvalidCommandException.java
@@ -1,3 +1,7 @@
+package lulu.exception;
+
+import lulu.exception.LuluException;
+
 public class InvalidCommandException extends LuluException {
     public InvalidCommandException() {
         super("(=✖ ᆺ ✖=) That's not a valid command.");
