+++ src/main/java/lulu/task/Task.java
@@ -1,3 +1,4 @@
+package lulu.task;
 public class Task {
     protected String description;
     protected boolean isDone;
