+++ src/main/java/lulu/task/Todo.java
@@ -1,3 +1,5 @@
+package lulu.task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
