+++ text-ui-test/EXPECTED.TXT
@@ -3,7 +3,7 @@ Hello! I am lulu (=◕ᆽ◕ฺ=)
 What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
-(=✖ ᆺ ✖=) That's not a valid command.
+(=✖ ᆺ ✖=) That's not a valid lulu.
 ____________________________________________________________
 ____________________________________________________________
 (=✖ ᆺ ✖=) The description of a todo cannot be empty.
